// src/bridge/ws/pb/code_pb.d.ts
export interface CodeMap {
  NONE: 0;
  MSG_D_REGISTER: 1;
  MSG_D_HEARTBEAT: 2;
// ...
  MSG_S_CACHE_DATA_UPDATE: 135;
  MSG_S_SUPPORT_MESSAGE: 136;
  MSG_S_USER_IS_MUTE: 137;
  MSG_S_USER_DIAMOND_CHANGE: 138;
  MSG_S_TRIBE_BLACK_USER_MTT: 139;
  MSG_S_ROOM_CHANGE_NOTIFY: 140;
  MSG_S_USER_GOLD_CHANGE: 141;
  MSG_S_TRIBE_BLACK_USER: 142;
  MSG_S_USER_TRADER_ORDER_NOTIFY: 143;
  MSG_S_ROOM_MTT_SETTLE_NOTIFY: 144;
  MSG_S_USER_USDT_ORDER_NOTIFY: 145;
  MSG_S_FRIEND_ROOM_CREATOR_SETTLE: 146;
  MSG_S_MTT_AWARD_NOTIFY: 147;
  MSG_S_USER_JOIN_CLUB: 148;
  MSG_S_CLUB_ROOM_MTT_SETTLE_NOTIFY: 149;
  MSG_S_USER_CLUB_ROLE_CHANGE: 150;
  MSG_S_USER_MTT_CHANGE_NOTIFY: 151;
  MSG_S_USER_SNG_CHANGE_NOTIFY: 152;
  MSG_S_MTT_SERIES_NOTIFY: 153;
  MSG_D_USER_PLAYING: 201;
  MSG_R_ROOMS: 202;
  MSG_R_QUICK_JOIN: 203;
  MSG_R_MTT_DETAIL: 204;
  MSG_R_JOIN_MATCHING: 205;
  MSG_R_ROOMS_SIMPLE: 206;
  MSG_S_UTIL_FACE_RECOGNIZE: 901;
  MSG_S_UTIL_ANTI_CHEAT_ROOM_VIDEO: 902;
  MSG_D_ENTER_ROOM: 1002;
  MSG_D_SEATED: 1003;
  MSG_D_ADD_ON: 1004;
  MSG_D_BRING_IN: 1005;
  MSG_D_ACTION: 1006;
  MSG_D_AUTO_OP_ACTIVE: 1007;
  MSG_D_SET_AUTO_ON_TABLE: 1008;
  MSG_D_STANDUP_ACTIVE: 1009;
  MSG_D_LEAVE: 1010;
// ...
  MSG_S_ACTION_ALL: 1108;
// ...
}

export const Code: CodeMap;